readline: move utilities to internal modules
