+++ lib/internal/console/constructor.js
@@ -429,7 +429,10 @@ const consoleMethods = {
     if (this._stdout.isTTY && process.env.TERM !== 'dumb') {
       // The require is here intentionally to avoid readline being
       // required too early when console is first loaded.
-      const { cursorTo, clearScreenDown } = require('readline');
+      const {
+        cursorTo,
+        clearScreenDown,
+      } = require('internal/readline/callbacks');
       cursorTo(this._stdout, 0, 0);
       clearScreenDown(this._stdout);
     }
