+++ lib/internal/repl/utils.js
@@ -41,7 +41,7 @@ const {
   clearScreenDown,
   cursorTo,
   moveCursor,
-} = require('readline');
+} = require('internal/readline/callbacks');
 
 const {
   commonPrefix,
